add ability to disable file writing in crash reports

+++ squiid-engine/src/crash_reporter.rs
@@ -37,7 +37,7 @@ impl fmt::Display for EnvironmentDetails<'_> {
     }
 }
 
-pub fn crash_report(panic_info: &PanicInfo) {
+pub fn crash_report(panic_info: &PanicInfo, write_dump_file: bool) {
     let backtrace = backtrace::Backtrace::new();
 
     // create environment struct
@@ -63,38 +63,40 @@ pub fn crash_report(panic_info: &PanicInfo) {
     );
 
     // determine the config directory to write the crash to
-    let result = panic::catch_unwind(config_handler::determine_config_path);
+    if write_dump_file {
+        let result = panic::catch_unwind(config_handler::determine_config_path);
 
-    let config_directory = match result {
-        Ok(value) => value.parent().map(|path| path.to_path_buf()),
-        Err(_) => match std::env::current_dir() {
-            Ok(value) => Some(value),
-            Err(_) => None,
-        },
-    };
+        let config_directory = match result {
+            Ok(value) => value.parent().map(|path| path.to_path_buf()),
+            Err(_) => match std::env::current_dir() {
+                Ok(value) => Some(value),
+                Err(_) => None,
+            },
+        };
 
-    if let Some(mut config_path_unwrapped) = config_directory {
-        config_path_unwrapped.push("squiid_crash.txt");
+        if let Some(mut config_path_unwrapped) = config_directory {
+            config_path_unwrapped.push("squiid_crash.txt");
 
-        // remove the old crash if it exists
-        let _ = std::fs::remove_file(&config_path_unwrapped);
+            // remove the old crash if it exists
+            let _ = std::fs::remove_file(&config_path_unwrapped);
 
-        let file = File::create(&config_path_unwrapped);
-        if let Ok(mut file_unwrapped) = file {
-            // write crash file
-            let crash_string = format!(
-                "Crash report generated at {}\n\n{}\n\n{}\n\n{:?}",
-                chrono::offset::Local::now(),
-                environment,
-                panic_info,
-                backtrace
-            );
-            let write_result = file_unwrapped.write_all(crash_string.as_bytes());
-            if write_result.is_ok() {
-                println!(
-                    "Crash report written to: {}",
-                    config_path_unwrapped.to_string_lossy()
+            let file = File::create(&config_path_unwrapped);
+            if let Ok(mut file_unwrapped) = file {
+                // write crash file
+                let crash_string = format!(
+                    "Crash report generated at {}\n\n{}\n\n{}\n\n{:?}",
+                    chrono::offset::Local::now(),
+                    environment,
+                    panic_info,
+                    backtrace
                 );
+                let write_result = file_unwrapped.write_all(crash_string.as_bytes());
+                if write_result.is_ok() {
+                    println!(
+                        "Crash report written to: {}",
+                        config_path_unwrapped.to_string_lossy()
+                    );
+                }
             }
         }
     }
